refactor audio: integrate partial changes from @humberthardy (#7) - support opus and mp3 audio based on browser support, preferrence for mp3 if supported - mp3 support much more stable, but not supported in Firefox, default to opus for firefox - move audio functions to audio.js, cleanup unnecessary code (@humberthardy + @ikreymer) - audio preference passed to init_browser sets AUDIO_TYPE (@humberthardy) - improve latency (between 200-500 ms of latency) @humberthardy - git-update-all.sh: check if git lfs is installed and init module for the first checkout (@humberthardy) - rename AUDIO_TYPE -> AUDIO_ALLOW, just boolean to enable/disable audio support. - update to new base-browser with mp3 support
