+++ docker-compose.yml
@@ -12,7 +12,7 @@ services:
             - CLUSTER_NAME=browser_shepherd
             - NETWORK_NAME=browsers_default
             - IDLE_TIMEOUT=60
-            - AUDIO_TYPE=opus
+            - AUDIO_ALLOWED=true
             - CONTAINER_EXPIRE_SECS=1200
 
         depends_on:
