+++ shepherd/static/browser.js
@@ -15,7 +15,6 @@ var CBrowser = function(reqid, target_div, init_params) {
     var vnc_pass = "secret";
 
     var end_time = undefined;
-    var cid = undefined;
 
     var waiting_for_container = false;
     var waiting_for_vnc = false;
@@ -42,8 +41,8 @@ var CBrowser = function(reqid, target_div, init_params) {
 
                     WebUtil.load_scripts(
                         {'core': ["base64.js", "websock.js", "des.js", "input/keysymdef.js",
-                                  "input/xtscancodes.js", "input/util.js", "input/devices.js",
-                                  "display.js", "inflator.js", "rfb.js", "input/keysym.js"]});
+                                "input/xtscancodes.js", "input/util.js", "input/devices.js",
+                                "display.js", "inflator.js", "rfb.js", "input/keysym.js"]});
                 });
             });
         }
@@ -69,7 +68,7 @@ var CBrowser = function(reqid, target_div, init_params) {
 
         $(init_params.clipboard).on('change keyup paste', function() {
             var text = $(init_params.clipboard).val();
-            
+
             if (connected && rfb && lastText != text) {
                 rfb.clipboardPasteFrom(text);
                 lastText = text;
@@ -144,7 +143,28 @@ var CBrowser = function(reqid, target_div, init_params) {
 
         req_params['reqid'] = reqid;
 
+        if (window.AUDIO_TYPES) {
+            add_audio(req_params, AUDIO_TYPES);
+        } else {
+            console.log("No Supported Audio Types");
+        }
+
         init_browser();
+    }
+
+    function add_audio(request, audio_types) {
+        delete req_params["audio"];
+
+        if (window.MediaSource) {
+            for (var i = 0; i < audio_types.length; i++) {
+                if (window.MediaSource.isTypeSupported(AUDIO_TYPES[i].type)) {
+                    req_params["audio"] = AUDIO_TYPES[i].id;
+                    return;
+                }
+            }
+        }
+
+        console.log("No Supported Audio Types Found");
     }
 
     function init_browser() {
@@ -189,8 +209,17 @@ var CBrowser = function(reqid, target_div, init_params) {
             vnc_pass = data.vnc_pass;
 
             if (init_params.audio) {
-                if (data.audio == "opus") {
-                    doAudio(data);
+                // setup_browser can be called many times (specially when noVnc thrown an exception), we stop sound before init again
+                if (window.hasOwnProperty("audioPlugin")) {
+                    try {
+                        window.audioPlugin.stop();
+                        window.audioPlugin = undefined;
+                    } catch (err){}
+
+                }
+                if (data.audio == "opus" || data.audio == "mp3") {
+                    window.audioPlugin = new WSAudio(data, init_params);
+                    window.audioPlugin.start();
                 }
             }
 
@@ -312,15 +341,15 @@ var CBrowser = function(reqid, target_div, init_params) {
 
         try {
             rfb = new RFB({'target':       canvas()[0],
-                           'encrypt':      (window.location.protocol === "https:"),
-                           'repeaterID':   '',
-                           'true_color':   true,
-                           'local_cursor': true,
-                           'shared':       false,
-                           'view_only':    false,
-                           'onUpdateState':  updateState,
-                           'onClipboard':    onVNCCopyCut,
-                           'onFBUComplete':  FBUComplete});
+                'encrypt':      (window.location.protocol === "https:"),
+                'repeaterID':   '',
+                'true_color':   true,
+                'local_cursor': true,
+                'shared':       false,
+                'view_only':    false,
+                'onUpdateState':  updateState,
+                'onClipboard':    onVNCCopyCut,
+                'onFBUComplete':  FBUComplete});
         } catch (exc) {
             waiting_for_vnc = false;
             //updateState(null, 'fatal', null, 'Unable to create RFB client -- ' + exc);
@@ -445,338 +474,14 @@ var CBrowser = function(reqid, target_div, init_params) {
                     if (init_params.on_event) {
                         init_params.on_event("reconnect");
                     }
- 
+
                     setup_browser();
                 }
             }
         });
     }
-
-    var audioCount = 0;
-
-    function get_ws_url(browser_info) {
-        var ws_url;
-
-        ws_url = (window.location.protocol == "https:" ? "wss:" : "ws:");
-
-        if (init_params.proxy_ws) {
-            var audio_port = browser_info.cmd_host.split(":")[1];
-            ws_url += window.location.host + "/" + init_params.proxy_ws + audio_port;
-            ws_url += "&count=" + audioCount++;
-        } else {
-            ws_url += browser_info.cmd_host + "/audio_ws";
-        }
-
-        return ws_url;
-    }
-
-    function get_http_url(browser_info) {
-        var audio_port = browser_info.cmd_host.split(":")[1];
-        var http_url = window.location.origin + "/_audio.webm?port=" + audio_port;
-        http_url += "&count=" + audioCount++;
-        return http_url;
-    }
-
-    var audioWS = null;
-
-    function WSAudio(ws_url) {
-        var MIME_TYPE = 'audio/webm; codecs="opus"';
-
-        var MAX_BUFFERS = 250;
-        var MIN_START_BUFFERS = 10;
-
-        var MAX_ERRORS = 1;
-
-        this.buffQ = [];
-        this.buffCount = 0;
-
-        this.ws_url = ws_url;
-        this.lastTs = undefined;
-        this.ws = null;
-
-        this.errCount = 0;
-
-        this.initing = false;
-        this.updating = false;
-
-        this.audio = null;
-        this.mediasource = null;
-        this.buffer = null;
-
-        this.buffSize = 0;
-
-        this.initAudio = function() {
-            if (this.initing) {
-                console.log("already initing");
-                return;
-            }
-
-            this.initing = true;
-            console.log("Init Audio");
-
-            this.mediasource = new MediaSource();
-            this.mediasource.addEventListener("sourceopen", this.sourceOpen.bind(this));
-
-            this.mediasource.addEventListener("error", (function(event) {
-                this.audioError("MediaSource Error", event);
-            }).bind(this));
-
-            this.audio = new Audio();
-            this.audio.src = URL.createObjectURL(this.mediasource);
-            this.audio.autoplay = true;
-            this.audio.load();
-            this.audio.play().catch(function() { });
-        }
-
-        this.sourceOpen = function() {
-            if (this.mediasource.sourceBuffers.length) {
-               console.log("source already open");
-               return;
-            }
-
-            var buffer = null;
-
-            try {
-                buffer = this.mediasource.addSourceBuffer(MIME_TYPE);
-            } catch (e) {
-                console.log("Opening Source Error: " + e);
-                return;
-            }
-
-            buffer.mode = "sequence";
-            //buffer.timestampOffset = 0;
-
-            buffer.addEventListener("error", (function(event) {
-                this.audioError("buffer error: " + (this.buffCount), event);
-            }).bind(this));
-
-            buffer.addEventListener("updateend", this.updateNext.bind(this));
-
-            this.buffer = buffer;
-            this.initing = false;
-
-            this.initWs();
-        };
-
-        this.close = function() {
-            console.log("Closing Audio");
-            try {
-                if (this.mediasource) {
-                    this.mediasource.removeSourceBuffer(this.buffer);
-                    if (this.mediasource.readyState == "open") {
-                        this.mediasource.endOfStream();
-                    }
-                }
-                this.buffer = null;
-
-            } catch(e) {
-                console.log("Error Closing: " + e);
-            }
-            this.mediasource = null;
-
-            try {
-                if (this.audio) {
-                    this.audio.pause();
-                }
-                this.audio = null;
-            } catch (e) {
-                console.log("Error Closing: " + e);
-            }
-
-            if (this.ws) {
-                try {
-                    this.ws.close();
-                } catch(e) {}
-
-                this.ws = null;
-            }
-        }
-
-        this.valid = function(lastTs) {
-            if (this.errCount > 0) {
-                return false;
-            }
-
-            if (!this.buffer) {
-                return false;
-            }
-
-            // if ws update hasn't changed
-            if (lastTs && this.lastTs && (this.lastTs <= lastTs)) {
-                return false;
-            }
-
-            return true;
-        }
-
-        this.mergeBuffers = function() {
-            var merged;
-
-            if (this.buffQ.length == 1) {
-                merged = this.buffQ[0];
-            } else {
-                merged = new Uint8Array(this.buffSize);
-
-                var length = this.buffQ.length;
-                var offset = 0;
-
-                for (var i = 0; i < length; i++) {
-                    var curr = this.buffQ[i];
-                    if (curr.length <= 0) {
-                        continue;
-                    }
-                    merged.set(curr, offset);
-                    offset += curr.length;
-                }
-            }
-
-            this.buffQ = [];
-            this.buffCount++;
-            return merged;
-        }
-
-        this.updateNext = function() {
-            if (this.initing || this.updating || (this.errCount >= MAX_ERRORS) || !this.buffQ.length || !this.buffer || this.buffer.updating) {
-                return;
-            }
-
-            if (this.buffCount == 0 && this.buffQ.length < MIN_START_BUFFERS) {
-                return;
-            }
-
-            this.updating = true;
-
-            try {
-                var merged = this.mergeBuffers();
-                this.buffer.appendBuffer(merged);
-                this.buffSize -= merged.length;
-                this.buffSize = 0;
-                this.errCount = 0;
-            } catch (e) {
-                this.audioError("Error Adding Buffer: " + e);
-            }
-
-            this.updating = false;
-        }
-
-        this.audioError = function(msg, event) {
-            if (this.audio.error) {
-                console.log(msg);
-                console.log(this.audio.error);
-                //console.log("num processed: " + this.buffCount);
-                this.errCount += 1;
-                if (this.errCount == 1) {
-                    setTimeout(window.restartAudio, 3000);
-                }
-            }
-        }
-
-        this.queue = function(buffer) {
-            if (this.buffSize > 20000) {
-                if (this.buffCount > 0) {
-                    res = this.buffQ.shift();
-                    this.buffSize -= res.length;
-                    console.log("dropping old buffers");
-                } else {
-                    // don't drop the starting buffers
-                    console.log("dropping new buffers");
-                    this.updateNext();
-                    return;
-                }
-            }
-
-            buffer = new Uint8Array(buffer);
-            this.buffQ.push(buffer);
-            this.buffSize += buffer.length;
-
-            this.updateNext();
-        }
-
-        this.initWs = function() {
-            try {
-                this.ws = new WebSocket(this.ws_url, "binary");
-            } catch (e) {
-                this.audioError("WS Open Failed");
-            }
-
-            this.ws.binaryType = 'arraybuffer';
-            this.ws.addEventListener("open", ws_open);
-            this.ws.addEventListener("close", ws_close);
-            this.ws.addEventListener("message", ws_message);
-            this.ws.addEventListener('error', ws_error);
-        }
-
-        var ws_open = function(event) {
-            if (!audioWS || this != audioWS.ws) {
-                return;
-            }
-
-            this.send("ready");
-        }
-
-        var ws_message = function(event) {
-            if (!audioWS || this != audioWS.ws) {
-                return;
-            }
-
-            if (audioWS.errCount < MAX_ERRORS) {
-                audioWS.queue(event.data);
-            }
-        }
-
-        var ws_error = function() {
-            if (!audioWS || this != audioWS.ws) {
-                return;
-            }
-
-            audioWS.audioError("WS Error");
-        }
-
-        var ws_close = function() {
-            if (!audioWS || this != audioWS.ws) {
-                return;
-            }
-
-            audioWS.audioError("WS Close");
-        }
-
-        this.initAudio();
-    }
-
-    var audioInit = false;
-    var browser_info;
-
-    function doAudio(info) {
-        if (audioInit) {
-            return;
-        }
-
-        browser_info = info;
-
-        audioInit = true;
-
-        window.restartAudio();
-    }
-
-    window.restartAudio = function() {
-        if (audioWS && audioWS.initing) {
-            return;
-        }
-
-        if (audioWS) {
-            console.log("Reiniting Audio");
-            audioWS.close();
-        }
-
-        audioWS = new WSAudio(get_ws_url(browser_info));
-    }
-
     start();
 
     return {"grab_focus": grab_focus,
             "lose_focus": lose_focus}
 };
-
-
-
-
