+++ shepherd/templates/browser_embed.html
@@ -3,7 +3,8 @@
 
 <head>
 <script src="/static/jquery-2.1.4.min.js"></script>
-<script src="/static/browser_controller.js"></script>
+<script src="/static/audio.js"></script>
+<script src="/static/browser.js"></script>
 
 <link href="/static/default.css" rel="stylesheet"/>
